relocate hLogin, hRegistration, hEditForm functions to minimize (not a function) bug
