+++ frontend/src/contexts/UserContext.jsx
@@ -5,20 +5,16 @@ import Query from "../services/Query";
 
 const UserContext = createContext({
   currentUser: {},
-  registrationForm: {},
-  setRegistrationForm: null,
-  userProfil: {},
-  setUserProfil: null,
-  setLoginForm: null,
-  loginForm: {},
-  hLogin: null,
+  videos: {},
+  favVideos: {},
+  isAuthenticated: 0,
+  hUserQueryRes: null,
 });
 export default UserContext;
 
 export function UserInfosContext({ children }) {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [videos, setVideos] = useState([]);
-  const [favVideos, setFavVideos] = useState([]);
   const [currentUser, setCurrentUser] = useState({
     username: "",
     lastname: "",
@@ -28,21 +24,6 @@ export function UserInfosContext({ children }) {
     address: "",
     firstname: "",
   });
-
-  const [registrationForm, setRegistrationForm] = useState({
-    username: "",
-    lastname: "",
-    email: "",
-    password: "",
-    firstname: "",
-    address: "",
-    city: "",
-    premium: 0,
-  });
-
-  const [loginForm, setLoginForm] = useState({ username: "", password: "" });
-
-  const [userProfil, setUserProfil] = useState(currentUser);
 
   useEffect(() => async () => setVideos(await Query.getAllVideos()), []);
 
@@ -63,77 +44,14 @@ export function UserInfosContext({ children }) {
     }
   };
 
-  const hLogin = (evt) => {
-    evt.preventDefault();
-    hUserQueryRes(Query.login(loginForm), "login");
-    swal({
-      title: "Bienvenue !",
-      text: "Bon visionnage !",
-      icon: "success",
-      buttons: false,
-      timer: 1000,
-    });
-  };
-
-  const hRegistration = (evt) => {
-    evt.preventDefault();
-    hUserQueryRes(Query.registration(registrationForm), "registration");
-    swal({
-      title: "Bienvenue !",
-      text: "Bon visionnage !",
-      icon: "success",
-      buttons: false,
-      timer: 1000,
-    });
-  };
-
-  const hEditFormSubmit = (evt) => {
-    evt.preventDefault();
-    hUserQueryRes(Query.editUser(userProfil, currentUser.id), "editing user");
-    swal({
-      title: "C'est noté !",
-      text: "Nous avons pris en compte vos changement !",
-      icon: "success",
-    });
-  };
-
-  useEffect(() => {
-    setUserProfil(currentUser);
-  }, [currentUser]);
-
   const context = useMemo(
     () => ({
       currentUser,
-      loginForm,
-      registrationForm,
-      userProfil,
       isAuthenticated,
       videos,
-      favVideos,
-      setUserProfil,
-      setRegistrationForm,
-      setCurrentUser,
-      setLoginForm,
-      setFavVideos,
-      setIsAuthenticated,
-      hLogin,
-      hRegistration,
-      hEditFormSubmit,
+      hUserQueryRes,
     }),
-    [
-      currentUser,
-      loginForm,
-      registrationForm,
-      userProfil,
-      isAuthenticated,
-      videos,
-      favVideos,
-      setUserProfil,
-      setRegistrationForm,
-      setLoginForm,
-      setCurrentUser,
-      setFavVideos,
-    ]
+    [currentUser, isAuthenticated, videos, hUserQueryRes]
   );
 
   return (
